fix: add forward declaration
