+++ src/creatures/players/wheel/wheel_gems.hpp
@@ -9,12 +9,19 @@
 
 #pragma once
 
-#include "wheel_definitions.hpp"
-#include "enums/player_wheel.hpp"
-
 class PlayerWheel;
 
+namespace WheelSpells {
+	struct Bonus;
+}
+
 enum CombatType_t : uint8_t;
+enum Vocation_t : uint16_t;
+
+enum class WheelGemAffinity_t : uint8_t;
+enum class WheelGemBasicModifier_t : uint8_t;
+enum class WheelGemSupremeModifier_t : uint8_t;
+enum class WheelStat_t : uint8_t;
 
 class GemModifierStrategy {
 public:
@@ -103,229 +110,9 @@ private:
 	Vocation_t m_vocation;
 };
 
-[[maybe_unused]] static int32_t getHealthValue(Vocation_t vocation, WheelGemBasicModifier_t modifier) {
-	static const std::unordered_map<WheelGemBasicModifier_t, std::unordered_map<Vocation_t, int32_t>> stats = {
-		{
-			WheelGemBasicModifier_t::Vocation_Health,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 300 },
-				{ Vocation_t::VOCATION_PALADIN, 200 },
-				{ Vocation_t::VOCATION_SORCERER, 100 },
-				{ Vocation_t::VOCATION_DRUID, 100 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Health_FireResistance,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 150 },
-				{ Vocation_t::VOCATION_PALADIN, 100 },
-				{ Vocation_t::VOCATION_SORCERER, 50 },
-				{ Vocation_t::VOCATION_DRUID, 50 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Health_EnergyResistance,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 150 },
-				{ Vocation_t::VOCATION_PALADIN, 100 },
-				{ Vocation_t::VOCATION_SORCERER, 50 },
-				{ Vocation_t::VOCATION_DRUID, 50 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Health_EarthResistance,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 150 },
-				{ Vocation_t::VOCATION_PALADIN, 100 },
-				{ Vocation_t::VOCATION_SORCERER, 50 },
-				{ Vocation_t::VOCATION_DRUID, 50 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Health_IceResistance,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 150 },
-				{ Vocation_t::VOCATION_PALADIN, 100 },
-				{ Vocation_t::VOCATION_SORCERER, 50 },
-				{ Vocation_t::VOCATION_DRUID, 50 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Mixed,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 150 },
-				{ Vocation_t::VOCATION_PALADIN, 100 },
-				{ Vocation_t::VOCATION_SORCERER, 50 },
-				{ Vocation_t::VOCATION_DRUID, 50 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Mixed2,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 150 },
-				{ Vocation_t::VOCATION_PALADIN, 100 },
-				{ Vocation_t::VOCATION_SORCERER, 50 },
-				{ Vocation_t::VOCATION_DRUID, 50 },
-			},
-		},
-	};
-
-	auto modifierIt = stats.find(modifier);
-	if (modifierIt != stats.end()) {
-		auto vocationIt = modifierIt->second.find(vocation);
-		if (vocationIt != modifierIt->second.end()) {
-			return vocationIt->second;
-		}
-	}
-	return 0;
-}
-
-[[maybe_unused]] static int32_t getManaValue(Vocation_t vocation, WheelGemBasicModifier_t modifier) {
-	static const std::unordered_map<WheelGemBasicModifier_t, std::unordered_map<Vocation_t, int32_t>> stats = {
-		{
-			WheelGemBasicModifier_t::Vocation_Mana_FireResistance,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 50 },
-				{ Vocation_t::VOCATION_PALADIN, 150 },
-				{ Vocation_t::VOCATION_SORCERER, 300 },
-				{ Vocation_t::VOCATION_DRUID, 300 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Mana_EnergyResistance,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 50 },
-				{ Vocation_t::VOCATION_PALADIN, 150 },
-				{ Vocation_t::VOCATION_SORCERER, 300 },
-				{ Vocation_t::VOCATION_DRUID, 300 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Mana_Earth_Resistance,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 50 },
-				{ Vocation_t::VOCATION_PALADIN, 150 },
-				{ Vocation_t::VOCATION_SORCERER, 300 },
-				{ Vocation_t::VOCATION_DRUID, 300 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Mana_Ice_Resistance,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 50 },
-				{ Vocation_t::VOCATION_PALADIN, 150 },
-				{ Vocation_t::VOCATION_SORCERER, 300 },
-				{ Vocation_t::VOCATION_DRUID, 300 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Mana,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 100 },
-				{ Vocation_t::VOCATION_PALADIN, 300 },
-				{ Vocation_t::VOCATION_SORCERER, 600 },
-				{ Vocation_t::VOCATION_DRUID, 600 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Mixed,
-			{
-				{ Vocation_t::VOCATION_PALADIN, 100 },
-				{ Vocation_t::VOCATION_SORCERER, 150 },
-				{ Vocation_t::VOCATION_DRUID, 150 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Capacity,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 50 },
-				{ Vocation_t::VOCATION_PALADIN, 150 },
-				{ Vocation_t::VOCATION_SORCERER, 300 },
-				{ Vocation_t::VOCATION_DRUID, 300 },
-			},
-		}
-	};
-
-	auto modifierIt = stats.find(modifier);
-	if (modifierIt != stats.end()) {
-		auto vocationIt = modifierIt->second.find(vocation);
-		if (vocationIt != modifierIt->second.end()) {
-			return vocationIt->second;
-		}
-	}
-	return 0;
-}
-
-[[maybe_unused]] static int32_t getCapacityValue(Vocation_t vocation, WheelGemBasicModifier_t modifier) {
-	static const std::unordered_map<WheelGemBasicModifier_t, std::unordered_map<Vocation_t, int32_t>> stats = {
-		{
-			WheelGemBasicModifier_t::Vocation_Capacity_FireResistance,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 250 },
-				{ Vocation_t::VOCATION_PALADIN, 200 },
-				{ Vocation_t::VOCATION_SORCERER, 100 },
-				{ Vocation_t::VOCATION_DRUID, 100 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Capacity_EnergyResistance,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 250 },
-				{ Vocation_t::VOCATION_PALADIN, 200 },
-				{ Vocation_t::VOCATION_SORCERER, 100 },
-				{ Vocation_t::VOCATION_DRUID, 100 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Capacity_EarthResistance,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 250 },
-				{ Vocation_t::VOCATION_PALADIN, 200 },
-				{ Vocation_t::VOCATION_SORCERER, 100 },
-				{ Vocation_t::VOCATION_DRUID, 100 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Capacity_IceResistance,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 250 },
-				{ Vocation_t::VOCATION_PALADIN, 200 },
-				{ Vocation_t::VOCATION_SORCERER, 100 },
-				{ Vocation_t::VOCATION_DRUID, 100 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Capacity,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 500 },
-				{ Vocation_t::VOCATION_PALADIN, 400 },
-				{ Vocation_t::VOCATION_SORCERER, 200 },
-				{ Vocation_t::VOCATION_DRUID, 200 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Mixed,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 125 },
-			},
-		},
-		{
-			WheelGemBasicModifier_t::Vocation_Mixed2,
-			{
-				{ Vocation_t::VOCATION_KNIGHT, 250 },
-				{ Vocation_t::VOCATION_PALADIN, 200 },
-				{ Vocation_t::VOCATION_SORCERER, 100 },
-				{ Vocation_t::VOCATION_DRUID, 100 },
-			},
-		}
-	};
-
-	auto modifierIt = stats.find(modifier);
-	if (modifierIt != stats.end()) {
-		auto vocationIt = modifierIt->second.find(vocation);
-		if (vocationIt != modifierIt->second.end()) {
-			return vocationIt->second;
-		}
-	}
-	return 0;
-}
+class WheelGemUtils {
+public:
+	[[maybe_unused]] static int32_t getHealthValue(Vocation_t vocation, WheelGemBasicModifier_t modifier);
+	[[maybe_unused]] static int32_t getManaValue(Vocation_t vocation, WheelGemBasicModifier_t modifier);
+	[[maybe_unused]] static int32_t getCapacityValue(Vocation_t vocation, WheelGemBasicModifier_t modifier);
+};
